[GPIOOUT]: Rename gpioOutput files.
