+++ core/ST/STM32L476/halLayer/gpio/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.25)
 
 add_library(GPIO_IMPLEMENTATION
         gpioPort.cpp
-        gpio.cpp
+        gpioOutput.cpp
 )
 
 target_link_libraries(GPIO_IMPLEMENTATION PRIVATE HAL_MCU_LAYER MCU_RELATED)
