+++ core/ST/STM32L476/halLayer/gpio/gpioOutput.h
@@ -3,8 +3,8 @@
 //
 #pragma once
 
-#include "IGpio.h"
 #include "stm32l476xx.h"
+#include "IGpioOutput.h"
 #include "components/gpio/IPort.h"
 #include "gpioPort.h"
